+++ pkg/authentication/scripts/check-benchmarks.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
 set -eu
 
-go test -run '^$' -bench Benchmark -benchmem -benchtime=100ms ./basic ./apikey ./authhttp
+go test -run '^$' -bench Benchmark -benchmem -benchtime=100ms ./basic ./apikey ./authhttp ./bearer
 (cd jwt && go test -run '^$' -bench Benchmark -benchmem -benchtime=100ms ./...)
 (cd oidc && go test -run '^$' -bench Benchmark -benchmem -benchtime=100ms ./...)